fiddle with the tracer scripts

+++ scripts/addtrace.sh
@@ -5,6 +5,13 @@ if ! [ -d ./tracer ]; then
 	exit 1
 fi
 
+if [ $(git status --porcelain | grep -v '^??' | wc -l) -ne 0 ]; then
+	echo "please clean up your working directory first (untracked files are okay)" >&2
+	exit 1
+fi
+
+echo "NOTE: use savetrace.sh to save updated tracers, use git reset --hard to get rid of the mess" >&2
+
 tmp=$(mktemp /tmp/droptrace.sh.XXXXXXXX)
 trap "rm -f '$tmp'" EXIT
 
@@ -24,4 +31,9 @@ find tracer -name 'tracer@*' | while read -r f; do
 	cat "$f" >>$tmp
 	cat "$file" | sed -n "$((lno+1)),\$p" >>$tmp
 	cat $tmp >$file
+done
+
+find libsrsirc platform src -name '*.c' | grep -vF -e 'base_log.c' -e 'helloworld.c' | while read -r f; do
+	echo "doing returns for $f" >&2
+	scripts/replret.sh "$f"
 done
